event_gateway/pyproject.toml
[tool.poetry]
name = "event-gateway"
version = "0.1.0"
description = "gateway for chat/auth events"
authors = ["ledxdeliveryflopp"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.115.6"}
sqlalchemy = "^2.0.36"
websockets = "^14.1"
psycopg2-binary = "^2.9.10"
asyncpg = "^0.30.0"
pydantic-settings = "^2.6.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
loguru = "^0.7.3"
aiofiles = "^24.1.0"
jinja2 = "^3.1.4"
httpx = "^0.28.1"
faststream = {extras = ["kafka"], version = "^0.5.33"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
